+++ backend/src/main/java/com/hoanghocdev/dolaspharmacy/service/impl/OrderServiceImpl.java
@@ -5,21 +5,22 @@ import com.hoanghocdev.dolaspharmacy.dto.request.OrderUpdateRequest;
 import com.hoanghocdev.dolaspharmacy.dto.response.OrderResponse;
 import com.hoanghocdev.dolaspharmacy.entity.Order;
 import com.hoanghocdev.dolaspharmacy.entity.OrderItem;
+import com.hoanghocdev.dolaspharmacy.entity.UserDetail;
 import com.hoanghocdev.dolaspharmacy.entity.Variant;
 import com.hoanghocdev.dolaspharmacy.entity.enums.OrderStatus;
 import com.hoanghocdev.dolaspharmacy.exception.AppException;
 import com.hoanghocdev.dolaspharmacy.exception.ErrorCode;
 import com.hoanghocdev.dolaspharmacy.mapper.OrderMapper;
-import com.hoanghocdev.dolaspharmacy.repository.CartRepository;
 import com.hoanghocdev.dolaspharmacy.repository.OrderRepository;
+import com.hoanghocdev.dolaspharmacy.repository.UserDetailRepository;
 import com.hoanghocdev.dolaspharmacy.repository.VariantRepository;
 import com.hoanghocdev.dolaspharmacy.service.CartService;
 import com.hoanghocdev.dolaspharmacy.service.OrderService;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,8 +32,8 @@ public class OrderServiceImpl implements OrderService {
     OrderMapper orderMapper;
     VariantRepository variantRepository;
     CartService cartService;
-    CartRepository cartRepository;
     OrderRepository orderRepository;
+    UserDetailRepository userDetailRepository;
 
     @Override
     public OrderResponse createOrder(OrderCreationRequest request) {
@@ -51,38 +52,50 @@ public class OrderServiceImpl implements OrderService {
         order.setOrderStatus(OrderStatus.PENDING);
 
         Order savedOrder = orderRepository.save(order);
-        return orderMapper.toOrderResponse(savedOrder);
+        return orderMapper.toResponse(savedOrder);
     }
 
     @Override
     public OrderResponse updateOrder(String orderId, OrderUpdateRequest request) {
         Order order = orderRepository.findById(orderId)
-                        .orElseThrow(() -> new AppException(ErrorCode.DATA_NOT_FOUND));
+                .orElseThrow(() -> new AppException(ErrorCode.DATA_NOT_FOUND));
         orderMapper.updateOrder(request, order);
-        if (request.getOrderStatus()!=null) {
+        if (request.getOrderStatus() != null) {
             order.setOrderStatus(OrderStatus.valueOf(request.getOrderStatus()));
         }
 
         order = orderRepository.save(order);
-        return orderMapper.toOrderResponse(order);
+        return orderMapper.toResponse(order);
     }
 
     @Override
-    public Page<OrderResponse> findOrderByPage(int page, int size) {
-        Pageable pageable = PageRequest.of(page, size);
+    public Page<OrderResponse> findOrderByPage(Pageable pageable) {
         return orderRepository.findAll(pageable)
-                .map(orderMapper::toOrderResponse);
+                .map(orderMapper::toResponse);
+    }
+
+    @Override
+    public Page<OrderResponse> findMyOrders(Pageable pageable) {
+        var contextHolder = SecurityContextHolder.getContext();
+        String name = contextHolder.getAuthentication().getName();
+        UserDetail userDetail = userDetailRepository.findByUsername(name)
+                .orElseThrow(() -> new AppException(ErrorCode.DATA_NOT_FOUND));
+        return orderRepository.findByUserDetail(userDetail, pageable)
+                .map(orderMapper::toResponse);
     }
 
     @Override
     public OrderResponse findOrderById(String orderId) {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new AppException(ErrorCode.DATA_NOT_FOUND));
-        return orderMapper.toOrderResponse(order);
+        return orderMapper.toResponse(order);
     }
 
     @Override
     public void deleteOrder(String orderId) {
+        if (!orderRepository.existsById(orderId)) {
+            throw new AppException(ErrorCode.DATA_NOT_FOUND);
+        }
         orderRepository.deleteById(orderId);
     }
 }
